+++ src/main/java/org/opentripplanner/openstreetmap/model/OSMNode.java
@@ -52,16 +52,6 @@ public class OSMNode extends OSMWithTags {
     }
 
     /**
-     * TODO Maybe all those methods (isSomething...) in OSMXxx should be moved to a dedicated OSM
-     * filtering class.
-     * 
-     * @return True if this node is a bike rental station.
-     */
-    public boolean isBikeRental() {
-        return isTag("amenity", "bicycle_rental");
-    }
-
-    /**
      * Checks if this node is bollard
      * @return true if it is
      */