+++ src/main/java/org/opentripplanner/routing/algorithm/astar/AStar.java
@@ -13,8 +13,6 @@ import org.opentripplanner.routing.graph.Vertex;
 import org.opentripplanner.routing.spt.GraphPath;
 import org.opentripplanner.routing.spt.ShortestPathTree;
 import org.opentripplanner.util.time.DateUtils;
-import org.opentripplanner.util.monitoring.MonitoringStore;
-import org.opentripplanner.util.monitoring.MonitoringStoreFactory;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -33,8 +31,6 @@ import java.util.List;
 public class AStar {
 
     private static final Logger LOG = LoggerFactory.getLogger(AStar.class);
-    // FIXME this is not really a factory, it's a way to fake a global variable. This should be stored at the OTPServer level.
-    private static final MonitoringStore store = MonitoringStoreFactory.getStore();
 
     private boolean verbose = false;
 
@@ -280,7 +276,6 @@ public class AStar {
             spt = runState.spt;
         }
         
-        storeMemory();
         return spt;
     }
     
@@ -306,15 +301,6 @@ public class AStar {
         }
         
         return spt;
-    }
-
-    private void storeMemory() {
-        if (store.isMonitoring("memoryUsed")) {
-            System.gc();
-            long memoryUsed = Runtime.getRuntime().totalMemory() -
-                    Runtime.getRuntime().freeMemory();
-            store.setLongMax("memoryUsed", memoryUsed);
-        }
     }
 
     public void setTraverseVisitor(TraverseVisitor traverseVisitor) {